refactor: create shared sql functions
