+++ packages/database/src/connectors/sqlite.ts
@@ -11,16 +11,17 @@ import {
   UpdateOptions,
   UpsertOptions,
   TableData,
-  normalizeRowDef,
+  // normalizeRowDef,
   constructSchema,
   Schema,
   Relation,
 } from '../types'
-
-const escapeQuotes = (str: string) => {
-  if (str === null) console.log(new Error().stack)
-  return str.replace(/"/gm, '""')
-}
+import {
+  // escapeQuotes,
+  whereToSql,
+  tableCreationSql,
+  parseType,
+} from '../helpers/sql'
 
 export class SQLiteConnector implements DB {
   db: any // Database<sqlite3.Database, sqlite3.Statement>
@@ -53,76 +54,6 @@ export class SQLiteConnector implements DB {
     const connector = new this(config)
     await connector.init()
     return connector
-  }
-
-  whereToSql(collection: string, doc: any = {}, sqlOnly = false) {
-    if (Object.keys(doc).length === 0) return ''
-    const table = this.schema[collection]
-    if (!table) throw new Error(`Unable to find table ${collection} in schema`)
-    const parseType = (type: string, value: any) => {
-      if (value === null) return 'NULL'
-      if (type === 'String') {
-        return `"${escapeQuotes(value)}"`
-      }
-      if (type === 'Int') {
-        return value
-      }
-      if (type === 'Bool') {
-        return value ? 'true' : 'false'
-      }
-      if (type === 'Object') {
-        return `"${escapeQuotes(JSON.stringify(value))}"`
-      }
-      throw new Error(`Unrecognized type ${type}`)
-    }
-    const sql = Object.keys(doc)
-      .map(key => {
-        if (key === 'OR') return
-        const rowDef = table.rows[key]
-        if (!rowDef)
-          throw new Error(`Unable to find row definition for key: "${key}"`)
-        const val = doc[key]
-        if (Array.isArray(val)) {
-          // need to generate an IN query
-          const values = val.map(v => parseType(rowDef.type, v))
-          return `"${key}" IN (${values.join(',')})`
-        }
-        if (typeof val === 'object' && val !== null) {
-          // parse lt, gt, lte, gte, ne operators
-          const operatorMap = {
-            lt: '<',
-            gt: '>',
-            lte: '<=',
-            gte: '>=',
-            ne: '!=',
-            eq: '=',
-          }
-          const nullOperatorMap = {
-            ne: 'IS NOT',
-            eq: 'IS',
-          }
-          return Object.keys(val).map(k => {
-            const operator =
-              val[k] === null ? nullOperatorMap[k] : operatorMap[k]
-            if (!operator) throw new Error(`Invalid operator ${k}`)
-            const parsed = parseType(rowDef.type, val[k])
-            return `"${key}" ${operator} ${parsed}`
-          })
-        }
-        const parsed = parseType(rowDef.type, val)
-        return `"${key}" ${parsed === 'NULL' ? 'IS' : '='} ${parsed}`
-      })
-      .flat()
-      .filter(i => !!i)
-      .join(' AND ')
-    if (Array.isArray(doc.OR)) {
-      const orConditions = doc.OR.map(w =>
-        this.whereToSql(collection, w, true),
-      ).join(' OR ')
-      return ` ${sqlOnly ? '' : 'WHERE'}
-      (${sql || 1}) AND (${orConditions})`
-    }
-    return ` ${sqlOnly ? '' : 'WHERE'} ${sql} `
   }
 
   async create(collection: string, _doc: any | any): Promise<any> {
@@ -181,26 +112,8 @@ export class SQLiteConnector implements DB {
           if (query[k]) {
             query[k].push(doc[k])
           }
-          const val = doc[k]
-          if (rowDef.type === 'Bool' && typeof val === 'boolean') {
-            return val ? 'true' : 'false'
-          }
-          if (rowDef.type === 'String' && typeof val === 'string') {
-            return `"${escapeQuotes(val)}"`
-          }
-          if (
-            rowDef.type === 'Int' &&
-            typeof val !== 'undefined' &&
-            val !== null
-          ) {
-            return `${val}`
-          }
-          if (rowDef.type === 'Object' && typeof val === 'object') {
-            return `"${escapeQuotes(JSON.stringify(val))}"`
-          }
-          return 'NULL'
+          return parseType(rowDef.type, doc[k])
         })
-        .filter(i => !!i)
         .join(',')
       allValues.push(`(${values})`)
     }
@@ -306,13 +219,13 @@ export class SQLiteConnector implements DB {
             .join(', ')}`
         : ''
     const limit = options.limit ? ` LIMIT ${options.limit} ` : ''
-    const sql = `SELECT * FROM "${collection}" ${this.whereToSql(
-      collection,
+    const table = this.schema[collection]
+    if (!table) throw new Error(`Unable to find table ${collection}`)
+    const sql = `SELECT * FROM "${collection}" ${whereToSql(
+      table,
       where,
     )} ${orderBy} ${limit};`
     const models = await this.db.all(sql)
-    const table = this.schema[collection]
-    if (!table) throw new Error(`Unable to find table ${collection}`)
     const objectKeys = Object.keys(table.rows).filter(key => {
       return table.rows[key]?.type === 'Object'
     })
@@ -342,8 +255,10 @@ export class SQLiteConnector implements DB {
   }
 
   async _count(collection: string, where: WhereClause) {
-    const sql = `SELECT COUNT(*) FROM "${collection}" ${this.whereToSql(
-      collection,
+    const table = this.schema[collection]
+    if (!table) throw new Error(`Unable to find table ${collection}`)
+    const sql = `SELECT COUNT(*) FROM "${collection}" ${whereToSql(
+      table,
       where,
     )};`
     const result = await this.db.get(sql)
@@ -366,28 +281,11 @@ export class SQLiteConnector implements DB {
         const rowDef = table.rows[key]
         if (!rowDef)
           throw new Error(`Unable to find row definition for key: "${key}"`)
-        const val = update[key]
-        if (val === null) {
-          return `"${key}" = NULL`
-        }
-        if (rowDef.type === 'String') {
-          return `"${key}" = "${escapeQuotes(val)}"`
-        }
-        if (rowDef.type === 'Int') {
-          return `"${key}" = ${val}`
-        }
-        if (rowDef.type === 'Bool') {
-          return `"${key}" = ${val ? 'true' : 'false'}`
-        }
-        if (rowDef.type === 'Object') {
-          return `"${key}" = "${escapeQuotes(JSON.stringify(val))}"`
-        }
-        throw new Error('Unknown row type')
+        return `"${key}" = ${parseType(rowDef.type, update[key])}`
       })
-      .filter(i => !!i)
       .join(', ')
-    const sql = `UPDATE "${collection}" SET ${setSql} ${this.whereToSql(
-      collection,
+    const sql = `UPDATE "${collection}" SET ${setSql} ${whereToSql(
+      table,
       where,
     )}`
     const result = await this.db.run(sql)
@@ -449,8 +347,8 @@ export class SQLiteConnector implements DB {
     const limitSql =
       options.limit === undefined ? '' : ` LIMIT ${options.limit} `
     const sql = `DELETE FROM "${collection}" WHERE "${table.primaryKey}" =
-    (SELECT "${table.primaryKey}" FROM "${collection}" ${this.whereToSql(
-      collection,
+    (SELECT "${table.primaryKey}" FROM "${collection}" ${whereToSql(
+      table,
       options.where,
     )} ${orderBySql} ${limitSql});`
     const { changes } = await this.db.run(sql)
@@ -468,50 +366,7 @@ export class SQLiteConnector implements DB {
 
   async createTables(tableData: TableData[]) {
     this.schema = constructSchema(tableData)
-    // run sql queries creating the tables as necessary
-    for (const table of tableData) {
-      const { name, primaryKey, rows } = table
-      const typeMap = {
-        String: 'TEXT',
-        Int: 'INTEGER',
-        Bool: 'BOOLEAN',
-        Object: 'TEXT', // serialize via json in connector
-      }
-      const rowCommands = rows
-        .map(row => {
-          const fullRow = normalizeRowDef(row)
-          // relations are virtual and assigned at load time
-          if (fullRow.relation) return
-          return `"${fullRow.name}" ${typeMap[fullRow.type]} ${
-            fullRow.optional ? '' : 'NOT NULL'
-          } ${fullRow.unique ? 'UNIQUE' : ''}`
-        })
-        .filter(i => !!i)
-      // Do i even need this if i'm loading manually????
-      const relationCommands = rows
-        .map(row => {
-          const fullRow = normalizeRowDef(row)
-          if (!fullRow.relation) return
-          return `FOREIGN KEY ("${fullRow.relation.localField}")
-          REFERENCES "${fullRow.relation.foreignTable}" ("${fullRow.relation.foreignField}")
-            ON DELETE SET NULL
-            ON UPDATE NO ACTION`
-        })
-        .filter(i => !!i)
-      if (primaryKey) {
-        const primaryKeys = [primaryKey]
-          .flat()
-          .map((name: string) => `"${name}"`)
-          .join(',')
-        relationCommands.push(`PRIMARY KEY (${primaryKeys})`)
-      }
-      // assume there's always at least 1 entry in rowCommands and relationCommands
-      const sql = `CREATE TABLE IF NOT EXISTS ${name} (
-        ${[rowCommands.join(','), relationCommands.join(',')]
-          .filter(i => !!i)
-          .join(',')}
-      );`
-      await this.db.exec(sql)
-    }
+    const createTablesCommand = tableCreationSql(tableData)
+    await this.db.exec(createTablesCommand)
   }
 }
